refactor: merge grpc+server, bstream:firehose->bstream:stream
note: disabled broken tests for now
